requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
idna==3.4
imageio==2.22.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-fid==0.2.1
pytz==2022.6
requests==2.28.1
scipy==1.9.3
six==1.16.0
torch==1.12.0+cu113
torchaudio==0.12.0+cu113
torchsummary==1.5.1
torchvision==0.13.0+cu113
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
